+++ src/Assembly/Metro/Controls/PageTemplates/Games/Components/MetaData/ColorData.cs
@@ -14,7 +14,7 @@ namespace Assembly.Metro.Controls.PageTemplates.Games.Components.MetaData
 		private Color _value;
 
 		public ColorData(string name, uint offset, long address, bool alpha, bool basic, string dataType, Color value,
-			uint pluginLine, string tooltip, Action<uint?, int> fieldSelected)
+			uint pluginLine, string tooltip, Action<long?, int> fieldSelected)
 			: base(name, offset, address, pluginLine, tooltip, fieldSelected)
 		{
 			_value = value;
