+++ src/Assembly/Metro/Controls/PageTemplates/Games/Components/MetaData/DataRef.cs
@@ -6,7 +6,7 @@ namespace Assembly.Metro.Controls.PageTemplates.Games.Components.MetaData
 	public class DataRef : RawData
 	{
 		public DataRef(string name, uint offset, string format, long address, long dataAddress, string value, int length,
-			uint pluginLine, string tooltip, Action<uint?, int> fieldSelected, FileSegmentGroup metaArea)
+			uint pluginLine, string tooltip, Action<long?, int> fieldSelected, FileSegmentGroup metaArea)
 			: base(name, offset, format, address, value, length, pluginLine, tooltip, fieldSelected, metaArea)
 		{
 			DataAddress = dataAddress;
