+++ src/Assembly/Metro/Controls/PageTemplates/Games/Components/MetaData/FlagData.cs
@@ -20,7 +20,7 @@ namespace Assembly.Metro.Controls.PageTemplates.Games.Components.MetaData
 		private FlagsType _type;
 		private ulong _value;
 
-		public FlagData(string name, uint offset, long address, FlagsType type, uint pluginLine, string tooltip, Action<uint?, int> fieldSelected)
+		public FlagData(string name, uint offset, long address, FlagsType type, uint pluginLine, string tooltip, Action<long?, int> fieldSelected)
 			: base(name, offset, address, pluginLine, tooltip, fieldSelected)
 		{
 			_type = type;
