+++ src/Assembly/Metro/Controls/PageTemplates/Games/Components/MetaData/MultiData.cs
@@ -7,7 +7,7 @@ namespace Assembly.Metro.Controls.PageTemplates.Games.Components.MetaData
 		private string _type;
 		internal T _a, _b;
 		
-		public Multi2Data(string name, uint offset, long address, string type, T a, T b, uint pluginLine, string tooltip, Action<uint?, int> fieldSelected)
+		public Multi2Data(string name, uint offset, long address, string type, T a, T b, uint pluginLine, string tooltip, Action<long?, int> fieldSelected)
 			: base(name, offset, address, pluginLine, tooltip, fieldSelected)
 		{
 			_type = type;
@@ -52,7 +52,7 @@ namespace Assembly.Metro.Controls.PageTemplates.Games.Components.MetaData
 		private string _type;
 		internal T _a, _b, _c;
 
-		public Multi3Data(string name, uint offset, long address, string type, T a, T b, T c, uint pluginLine, string tooltip, Action<uint?, int> fieldSelected)
+		public Multi3Data(string name, uint offset, long address, string type, T a, T b, T c, uint pluginLine, string tooltip, Action<long?, int> fieldSelected)
 			: base(name, offset, address, pluginLine, tooltip, fieldSelected)
 		{
 			_type = type;
@@ -108,7 +108,7 @@ namespace Assembly.Metro.Controls.PageTemplates.Games.Components.MetaData
 		private string _type;
 		internal T _a, _b, _c, _d;
 
-		public Multi4Data(string name, uint offset, long address, string type, T a, T b, T c, T d, uint pluginLine, string tooltip, Action<uint?, int> fieldSelected)
+		public Multi4Data(string name, uint offset, long address, string type, T a, T b, T c, T d, uint pluginLine, string tooltip, Action<long?, int> fieldSelected)
 			: base(name, offset, address, pluginLine, tooltip, fieldSelected)
 		{
 			_type = type;
@@ -172,7 +172,7 @@ namespace Assembly.Metro.Controls.PageTemplates.Games.Components.MetaData
 
 	public class Degree2Data : Multi2Data<float>
 	{
-		public Degree2Data(string name, uint offset, long address, string type, float a, float b, uint pluginLine, string tooltip, Action<uint?, int> fieldSelected)
+		public Degree2Data(string name, uint offset, long address, string type, float a, float b, uint pluginLine, string tooltip, Action<long?, int> fieldSelected)
 			: base(name, offset, address, type, a, b, pluginLine, tooltip, fieldSelected)
 		{
 		}
@@ -234,7 +234,7 @@ namespace Assembly.Metro.Controls.PageTemplates.Games.Components.MetaData
 
 	public class Degree3Data : Multi3Data<float>
 	{
-		public Degree3Data(string name, uint offset, long address, string type, float a, float b, float c, uint pluginLine, string tooltip, Action<uint?, int> fieldSelected)
+		public Degree3Data(string name, uint offset, long address, string type, float a, float b, float c, uint pluginLine, string tooltip, Action<long?, int> fieldSelected)
 			: base(name, offset, address, type, a, b, c, pluginLine, tooltip, fieldSelected)
 		{
 		}
@@ -317,7 +317,7 @@ namespace Assembly.Metro.Controls.PageTemplates.Games.Components.MetaData
 
 	public class Vector2Data : Multi2Data<float>
 	{
-		public Vector2Data(string name, uint offset, long address, string type, float a, float b, uint pluginLine, string tooltip, Action<uint?, int> fieldSelected)
+		public Vector2Data(string name, uint offset, long address, string type, float a, float b, uint pluginLine, string tooltip, Action<long?, int> fieldSelected)
 			: base(name, offset, address, type, a, b, pluginLine, tooltip, fieldSelected)
 		{
 		}
@@ -337,7 +337,7 @@ namespace Assembly.Metro.Controls.PageTemplates.Games.Components.MetaData
 
 	public class Vector3Data : Multi3Data<float>
 	{
-		public Vector3Data(string name, uint offset, long address, string type, float a, float b, float c, uint pluginLine, string tooltip, Action<uint?, int> fieldSelected)
+		public Vector3Data(string name, uint offset, long address, string type, float a, float b, float c, uint pluginLine, string tooltip, Action<long?, int> fieldSelected)
 			: base(name, offset, address, type, a, b, c, pluginLine, tooltip, fieldSelected)
 		{
 		}
@@ -357,7 +357,7 @@ namespace Assembly.Metro.Controls.PageTemplates.Games.Components.MetaData
 
 	public class Vector4Data : Multi4Data<float>
 	{
-		public Vector4Data(string name, uint offset, long address, string type, float a, float b, float c, float d, uint pluginLine, string tooltip, Action<uint?, int> fieldSelected)
+		public Vector4Data(string name, uint offset, long address, string type, float a, float b, float c, float d, uint pluginLine, string tooltip, Action<long?, int> fieldSelected)
 			: base(name, offset, address, type, a, b, c, d, pluginLine, tooltip, fieldSelected)
 		{
 		}
@@ -377,7 +377,7 @@ namespace Assembly.Metro.Controls.PageTemplates.Games.Components.MetaData
 
 	public class Point2Data : Multi2Data<float>
 	{
-		public Point2Data(string name, uint offset, long address, string type, float a, float b, uint pluginLine, string tooltip, Action<uint?, int> fieldSelected)
+		public Point2Data(string name, uint offset, long address, string type, float a, float b, uint pluginLine, string tooltip, Action<long?, int> fieldSelected)
 			: base(name, offset, address, type, a, b, pluginLine, tooltip, fieldSelected)
 		{
 		}
@@ -396,7 +396,7 @@ namespace Assembly.Metro.Controls.PageTemplates.Games.Components.MetaData
 
 	public class Point3Data : Multi3Data<float>
 	{
-		public Point3Data(string name, uint offset, long address, string type, float a, float b, float c, uint pluginLine, string tooltip, Action<uint?, int> fieldSelected)
+		public Point3Data(string name, uint offset, long address, string type, float a, float b, float c, uint pluginLine, string tooltip, Action<long?, int> fieldSelected)
 			: base(name, offset, address, type, a, b, c, pluginLine, tooltip, fieldSelected)
 		{
 		}
@@ -416,7 +416,7 @@ namespace Assembly.Metro.Controls.PageTemplates.Games.Components.MetaData
 
 	public class Plane2Data : Multi3Data<float>
 	{
-		public Plane2Data(string name, uint offset, long address, string type, float a, float b, float c, uint pluginLine, string tooltip, Action<uint?, int> fieldSelected)
+		public Plane2Data(string name, uint offset, long address, string type, float a, float b, float c, uint pluginLine, string tooltip, Action<long?, int> fieldSelected)
 			: base(name, offset, address, type, a, b, c, pluginLine, tooltip, fieldSelected)
 		{
 		}
@@ -436,7 +436,7 @@ namespace Assembly.Metro.Controls.PageTemplates.Games.Components.MetaData
 
 	public class Plane3Data : Multi4Data<float>
 	{
-		public Plane3Data(string name, uint offset, long address, string type, float a, float b, float c, float d, uint pluginLine, string tooltip, Action<uint?, int> fieldSelected)
+		public Plane3Data(string name, uint offset, long address, string type, float a, float b, float c, float d, uint pluginLine, string tooltip, Action<long?, int> fieldSelected)
 			: base(name, offset, address, type, a, b, c, d, pluginLine, tooltip, fieldSelected)
 		{
 		}
@@ -456,7 +456,7 @@ namespace Assembly.Metro.Controls.PageTemplates.Games.Components.MetaData
 
 	public class RectangleData : Multi4Data<short>
 	{
-		public RectangleData(string name, uint offset, long address, string type, short a, short b, short c, short d, uint pluginLine, string tooltip, Action<uint?, int> fieldSelected)
+		public RectangleData(string name, uint offset, long address, string type, short a, short b, short c, short d, uint pluginLine, string tooltip, Action<long?, int> fieldSelected)
 			: base(name, offset, address, type, a, b, c, d, pluginLine, tooltip, fieldSelected)
 		{
 		}
@@ -476,7 +476,7 @@ namespace Assembly.Metro.Controls.PageTemplates.Games.Components.MetaData
 
 	public class Quaternion16Data : Multi4Data<short>
 	{
-		public Quaternion16Data(string name, uint offset, long address, string type, short a, short b, short c, short d, uint pluginLine, string tooltip, Action<uint?, int> fieldSelected)
+		public Quaternion16Data(string name, uint offset, long address, string type, short a, short b, short c, short d, uint pluginLine, string tooltip, Action<long?, int> fieldSelected)
 			: base(name, offset, address, type, a, b, c, d, pluginLine, tooltip, fieldSelected)
 		{
 		}
@@ -496,7 +496,7 @@ namespace Assembly.Metro.Controls.PageTemplates.Games.Components.MetaData
 
 	public class Point16Data : Multi2Data<short>
 	{
-		public Point16Data(string name, uint offset, long address, string type, short a, short b, uint pluginLine, string tooltip, Action<uint?, int> fieldSelected)
+		public Point16Data(string name, uint offset, long address, string type, short a, short b, uint pluginLine, string tooltip, Action<long?, int> fieldSelected)
 			: base(name, offset, address, type, a, b, pluginLine, tooltip, fieldSelected)
 		{
 		}
