+++ src/Assembly/Metro/Controls/PageTemplates/Games/Components/MetaData/StringIDData.cs
@@ -8,7 +8,7 @@ namespace Assembly.Metro.Controls.PageTemplates.Games.Components.MetaData
 		private Trie _autocompleteTrie;
 		private string _value;
 
-		public StringIDData(string name, uint offset, long address, string val, Trie autocompleteTrie, uint pluginLine, string tooltip, Action<uint?, int> fieldSelected)
+		public StringIDData(string name, uint offset, long address, string val, Trie autocompleteTrie, uint pluginLine, string tooltip, Action<long?, int> fieldSelected)
 			: base(name, offset, address, pluginLine, tooltip, fieldSelected)
 		{
 			_value = val;
