+++ src/Assembly/Metro/Controls/PageTemplates/Games/Components/MetaData/TagRefData.cs
@@ -12,7 +12,7 @@ namespace Assembly.Metro.Controls.PageTemplates.Games.Components.MetaData
 		private TagEntry _value;
 
 		public TagRefData(string name, uint offset, long address, TagHierarchy allTags, Visibility showTagOptions, bool withGroup,
-			uint pluginLine, string tooltip, Action<uint?, int> fieldSelected)
+			uint pluginLine, string tooltip, Action<long?, int> fieldSelected)
 			: base(name, offset, address, pluginLine, tooltip, fieldSelected)
 		{
 			_allTags = allTags;
